# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NexaShift AI Earnings Command Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /policy/create:
    post:
      operationId: createPolicy
      tags: [policy]
      summary: Create a new income protection policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePolicyBody"
      responses:
        "201":
          description: Policy created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
  /policy/{userId}:
    get:
      operationId: getUserPolicy
      tags: [policy]
      summary: Get user's active policy
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
  /claim/process:
    post:
      operationId: processClaim
      tags: [claims]
      summary: Auto-process a claim (zero-touch)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimRequestBody"
      responses:
        "200":
          description: Claim processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimResult"
  /claims/{userId}:
    get:
      operationId: getUserClaims
      tags: [claims]
      summary: Get claim history for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Claims list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Claim"
...
  /daily-plan/{userId}:
    get:
      operationId: getDailyPlan
      tags: [decisions]
      summary: Get AI-generated daily work plan
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Daily plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyPlan"
...
  /insights/{userId}:
    get:
      operationId: getUserInsights
      tags: [insights]
      summary: Get behavioral insights for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User insights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInsights"
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [insights]
      summary: Get aggregated dashboard summary for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
...
components:
  schemas:
...
    CreatePolicyBody:
      type: object
      properties:
        userId:
          type: string
        coverageType:
          type: string
          enum: [basic, standard, premium]
      required:
        - userId
        - coverageType
    Policy:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        coverageType:
          type: string
        premium:
          type: number
        coverage:
          type: number
        status:
          type: string
          enum: [active, inactive, suspended]
        startDate:
          type: string
        nextPaymentDate:
          type: string
        claimsCount:
          type: integer
        totalPaidOut:
          type: number
      required:
        - id
        - userId
        - coverageType
        - premium
        - coverage
        - status
        - startDate
        - nextPaymentDate
        - claimsCount
        - totalPaidOut
    ClaimRequestBody:
      type: object
      properties:
        userId:
          type: string
        triggerType:
          type: string
          enum: [rain, aqi, heat, platform_outage, upi_downtime]
        incomeImpact:
          type: number
          description: Estimated income loss in INR
      required:
        - userId
        - triggerType
        - incomeImpact
    ClaimResult:
      type: object
      properties:
        claimId:
          type: string
        status:
          type: string
          enum: [approved, rejected, under_review]
        payout:
          type: number
        fraudScore:
          type: number
        trustScore:
          type: number
        explanation:
          type: string
        processingTime:
          type: string
      required:
        - claimId
        - status
        - payout
        - fraudScore
        - trustScore
        - explanation
        - processingTime
    Claim:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        triggerType:
          type: string
        incomeImpact:
          type: number
        payout:
          type: number
        status:
          type: string
        fraudScore:
          type: number
        explanation:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - triggerType
        - incomeImpact
        - payout
        - status
        - fraudScore
        - explanation
        - createdAt
...
    DailyPlan:
      type: object
      properties:
        userId:
          type: string
        date:
          type: string
        bestWorkingHours:
          type: array
          items:
            $ref: "#/components/schemas/WorkWindow"
        highDemandZones:
          type: array
          items:
            $ref: "#/components/schemas/Zone"
        riskWindows:
          type: array
          items:
            $ref: "#/components/schemas/RiskWindow"
        expectedEarnings:
          type: number
        recommendation:
          type: string
        weatherForecast:
          type: string
        topTip:
          type: string
      required:
        - userId
        - date
        - bestWorkingHours
        - highDemandZones
        - riskWindows
        - expectedEarnings
        - recommendation
        - weatherForecast
        - topTip
    WorkWindow:
      type: object
      properties:
        startTime:
          type: string
        endTime:
          type: string
        demandScore:
          type: number
        estimatedEarning:
          type: number
        label:
          type: string
      required:
        - startTime
        - endTime
        - demandScore
        - estimatedEarning
        - label
    Zone:
      type: object
      properties:
        name:
          type: string
        demandScore:
          type: number
        avgEarning:
          type: number
        distanceKm:
          type: number
      required:
        - name
        - demandScore
        - avgEarning
        - distanceKm
    RiskWindow:
      type: object
      properties:
        startTime:
          type: string
        endTime:
          type: string
        riskType:
          type: string
        severity:
          type: string
        advice:
          type: string
      required:
        - startTime
        - endTime
        - riskType
        - severity
        - advice
...
    UserInsights:
      type: object
      properties:
        userId:
          type: string
        bestEarningTime:
          type: string
        bestEarningDay:
          type: string
        riskExposure:
          type: number
        performanceScore:
          type: number
        weeklyEarnings:
          type: array
          items:
            $ref: "#/components/schemas/DailyEarning"
        monthlyTrend:
          type: string
          enum: [improving, stable, declining]
        topRiskFactors:
          type: array
          items:
            type: string
        savingsRecommendation:
          type: number
      required:
        - userId
        - bestEarningTime
        - bestEarningDay
        - riskExposure
        - performanceScore
        - weeklyEarnings
        - monthlyTrend
        - topRiskFactors
        - savingsRecommendation
    DailyEarning:
      type: object
      properties:
        day:
          type: string
        amount:
          type: number
        hoursWorked:
          type: number
      required:
        - day
        - amount
        - hoursWorked
...
    DashboardSummary:
      type: object
      properties:
        userId:
          type: string
        todayEarnings:
          type: number
        weeklyEarnings:
          type: number
        monthlyEarnings:
          type: number
        currentRiskScore:
          type: number
        activePolicies:
          type: integer
        totalClaimsThisMonth:
          type: integer
        totalPayoutThisMonth:
          type: number
        riskTrend:
          type: string
          enum: [improving, stable, worsening]
        incomeProtected:
          type: number
        streakDays:
          type: integer
      required:
        - userId
        - todayEarnings
        - weeklyEarnings
        - monthlyEarnings
        - currentRiskScore
        - activePolicies
        - totalClaimsThisMonth
        - totalPayoutThisMonth
        - riskTrend
        - incomeProtected
        - streakDays