# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NexaShift AI Earnings Command Center API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new gig worker
 */
export const RegisterUserBody = zod.object({
  name: zod.string(),
  city: zod.string(),
  income: zod.number().describe("Monthly income in INR"),
  workType: zod.enum([
    "delivery",
    "rideshare",
    "freelance",
    "domestic_help",
    "construction",
  ]),
});

/**
 * @summary Create a new income protection policy
 */
export const CreatePolicyBody = zod.object({
  userId: zod.string(),
  coverageType: zod.enum(["basic", "standard", "premium"]),
});

/**
 * @summary Get user's active policy
 */
export const GetUserPolicyParams = zod.object({
  userId: zod.coerce.string(),
});

export const GetUserPolicyResponse = zod.object({
  id: zod.string(),
  userId: zod.string(),
  coverageType: zod.string(),
  premium: zod.number(),
  coverage: zod.number(),
  status: zod.enum(["active", "inactive", "suspended"]),
  startDate: zod.string(),
  nextPaymentDate: zod.string(),
  claimsCount: zod.number(),
  totalPaidOut: zod.number(),
});

/**
 * @summary Auto-process a claim (zero-touch)
 */
export const ProcessClaimBody = zod.object({
  userId: zod.string(),
  triggerType: zod.enum([
    "rain",
    "aqi",
    "heat",
    "platform_outage",
    "upi_downtime",
  ]),
  incomeImpact: zod.number().describe("Estimated income loss in INR"),
});

export const ProcessClaimResponse = zod.object({
  claimId: zod.string(),
  status: zod.enum(["approved", "rejected", "under_review"]),
  payout: zod.number(),
  fraudScore: zod.number(),
  trustScore: zod.number(),
  explanation: zod.string(),
  processingTime: zod.string(),
});

/**
 * @summary Get claim history for a user
 */
export const GetUserClaimsParams = zod.object({
  userId: zod.coerce.string(),
});

export const GetUserClaimsResponseItem = zod.object({
  id: zod.string(),
  userId: zod.string(),
  triggerType: zod.string(),
  incomeImpact: zod.number(),
  payout: zod.number(),
  status: zod.string(),
  fraudScore: zod.number(),
  explanation: zod.string(),
  createdAt: zod.string(),
});
export const GetUserClaimsResponse = zod.array(GetUserClaimsResponseItem);

/**
 * @summary Get active environmental triggers
 */
export const GetTriggersResponse = zod.object({
  activeTriggers: zod.array(
    zod.object({
      type: zod.enum([
        "rain",
        "aqi",
        "heat",
        "platform_outage",
        "upi_downtime",
      ]),
      severity: zod.enum(["low", "medium", "high", "critical"]),
      description: zod.string(),
      incomeImpactPercent: zod.number(),
      cities: zod.array(zod.string()),
      isActive: zod.boolean(),
    }),
  ),
  riskLevel: zod.enum(["low", "medium", "high", "critical"]),
  affectedCities: zod.array(zod.string()),
});

/**
 * @summary What-if simulation engine
 */
export const SimulateScenarioQueryParams = zod.object({
  scenario: zod.enum([
    "rain",
    "location_change",
    "skip_work",
    "platform_outage",
    "aqi_spike",
    "heat_wave",
    "upi_downtime",
  ]),
  userId: zod.coerce.string().optional(),
});

export const SimulateScenarioResponse = zod.object({
  scenario: zod.string(),
  predictedIncome: zod.number(),
  baseIncome: zod.number(),
  riskPercent: zod.number(),
  incomeDrop: zod.number(),
  incomeDropPercent: zod.number(),
  recommendedAction: zod.string(),
  expectedPayout: zod.number(),
  explanation: zod.string(),
  confidence: zod.number().describe("0-100 confidence score"),
  alternativeScenarios: zod.array(
    zod.object({
      name: zod.string(),
      incomeChange: zod.number(),
      action: zod.string(),
    }),
  ),
});

/**
 * @summary Get AI-generated daily work plan
 */
export const GetDailyPlanParams = zod.object({
  userId: zod.coerce.string(),
});

export const GetDailyPlanResponse = zod.object({
  userId: zod.string(),
  date: zod.string(),
  bestWorkingHours: zod.array(
    zod.object({
      startTime: zod.string(),
      endTime: zod.string(),
      demandScore: zod.number(),
      estimatedEarning: zod.number(),
      label: zod.string(),
    }),
  ),
  highDemandZones: zod.array(
    zod.object({
      name: zod.string(),
      demandScore: zod.number(),
      avgEarning: zod.number(),
      distanceKm: zod.number(),
    }),
  ),
  riskWindows: zod.array(
    zod.object({
      startTime: zod.string(),
      endTime: zod.string(),
      riskType: zod.string(),
      severity: zod.string(),
      advice: zod.string(),
    }),
  ),
  expectedEarnings: zod.number(),
  recommendation: zod.string(),
  weatherForecast: zod.string(),
  topTip: zod.string(),
});

/**
 * @summary Get city-level map data with risk and demand
 */
export const GetMapDataResponseItem = zod.object({
  city: zod.string(),
  lat: zod.number(),
  lng: zod.number(),
  riskScore: zod.number(),
  demandScore: zod.number(),
  avgPayout: zod.number(),
  activeClaims: zod.number(),
  weather: zod.string(),
  aqiLevel: zod.string(),
});
export const GetMapDataResponse = zod.array(GetMapDataResponseItem);

/**
 * @summary Get behavioral insights for a user
 */
export const GetUserInsightsParams = zod.object({
  userId: zod.coerce.string(),
});

export const GetUserInsightsResponse = zod.object({
  userId: zod.string(),
  bestEarningTime: zod.string(),
  bestEarningDay: zod.string(),
  riskExposure: zod.number(),
  performanceScore: zod.number(),
  weeklyEarnings: zod.array(
    zod.object({
      day: zod.string(),
      amount: zod.number(),
      hoursWorked: zod.number(),
    }),
  ),
  monthlyTrend: zod.enum(["improving", "stable", "declining"]),
  topRiskFactors: zod.array(zod.string()),
  savingsRecommendation: zod.number(),
});

/**
 * @summary Get AI explanation for a prediction
 */
export const GetAiExplanationQueryParams = zod.object({
  event: zod.coerce.string(),
  userId: zod.coerce.string().optional(),
});

export const GetAiExplanationResponse = zod.object({
  event: zod.string(),
  factors: zod.array(
    zod.object({
      factor: zod.string(),
      impact: zod.string(),
      weight: zod.number(),
      description: zod.string(),
    }),
  ),
  summary: zod.string(),
  confidence: zod.number(),
  historicalAccuracy: zod.number(),
});

/**
 * @summary Get aggregated dashboard summary for a user
 */
export const GetDashboardSummaryQueryParams = zod.object({
  userId: zod.coerce.string(),
});

export const GetDashboardSummaryResponse = zod.object({
  userId: zod.string(),
  todayEarnings: zod.number(),
  weeklyEarnings: zod.number(),
  monthlyEarnings: zod.number(),
  currentRiskScore: zod.number(),
  activePolicies: zod.number(),
  totalClaimsThisMonth: zod.number(),
  totalPayoutThisMonth: zod.number(),
  riskTrend: zod.enum(["improving", "stable", "worsening"]),
  incomeProtected: zod.number(),
  streakDays: zod.number(),
});

/**
 * @summary List all registered users
 */
export const ListUsersResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  city: zod.string(),
  income: zod.number(),
  workType: zod.string(),
  riskScore: zod.number(),
  createdAt: zod.string(),
});
export const ListUsersResponse = zod.array(ListUsersResponseItem);